+++ android/src/main/kotlin/com/indiegeeker/flutter_app_updater/FlutterAppUpdaterPlugin.kt
@@ -92,7 +92,6 @@ class FlutterAppUpdaterPlugin private constructor(
       }
       "getDownloadPath" -> getDownloadPath(result)
       "openStore" -> openStore(call, result)
-      "startPlayInAppUpdate" -> startPlayInAppUpdate(result)
       "openAndroidMarket" -> openAndroidMarket(call, result)
       else -> result.notImplemented()
     }
@@ -297,14 +296,6 @@ class FlutterAppUpdaterPlugin private constructor(
     } catch (e: Exception) {
       false
     }
-  }
-
-  private fun startPlayInAppUpdate(result: Result) {
-    result.error(
-      "PLAY_IN_APP_UPDATE_UNAVAILABLE",
-      "当前原生插件未集成Google Play In-App Updates",
-      null
-    )
   }
 
   private fun openAndroidMarket(call: MethodCall, result: Result) {