+++ lib/src/channel/flutter_app_updater_platform_interface.dart
@@ -52,14 +52,6 @@ abstract class FlutterAppUpdaterPlatform extends PlatformInterface {
     throw UnimplementedError('openStore() has not been implemented.');
   }
 
-  Future<void> startPlayInAppUpdate({
-    required String mode,
-  }) {
-    throw UnimplementedError(
-      'startPlayInAppUpdate() has not been implemented.',
-    );
-  }
-
   Future<void> openAndroidMarket({
     required String marketPackageName,
     required String marketUri,