+++ lib/src/models/update_error_code.dart
@@ -10,7 +10,6 @@ enum UpdateErrorCode {
   noSupportedAction('NO_SUPPORTED_ACTION'),
   storeNotAvailable('STORE_NOT_AVAILABLE'),
   marketNotAvailable('MARKET_NOT_AVAILABLE'),
-  playInAppUpdateUnavailable('PLAY_IN_APP_UPDATE_UNAVAILABLE'),
   packageDownloadFailed('PACKAGE_DOWNLOAD_FAILED'),
   packageTooLarge('PACKAGE_TOO_LARGE'),
   packageHashMismatch('PACKAGE_HASH_MISMATCH'),